use deviceId in the event name
reason: It is the emitter role to split message to the right node,
not the node role. If we have many hundreds of node, then it can have
impact on performance

If on day, we want to support dynamic device ID, then I guess we could
send two different events (one with ID and another generic). But for now
we don't support it and don't plan to support it