retrieve credentials from editor API
credentials are not part of the flow, so there are not preloaded like
nodes. We need to fetch them dynamically if there are not loaded. Also
to avoid desync when we edit the config node without saving it, we need
to store the credentials in the same attribute as the core code (e.i.
node.credentials) and it require to follow the same logic (e.i. populate
the node.credentials._) to avoid breaking the concept of dirty node.

It result to a weak design, if node-red code change how it handled
credentials, then it may break nodes too. Anyways, it is the same thing
for RED.nodes.node.