device: set name dynamically according the label
The name was hardcoded on the first save of the node. But it can be
really annoying if you change your device name. With this modification,
you can leave the name empty and the node will track the device name
each time you save it.

limitation: if you change your device name but doesn't click to edit the
node, the node will remains with the old name until you open node and
click on "Done" to save it